+++ common-helpers.js
@@ -395,6 +395,20 @@ export function chunkString(str, length) {
     return str.match(new RegExp(".{1," + length + "}", "g"));
 }
 
+/**
+ * Converts a data URL to a Blob
+ * @param {string} dataurl - The data URL to convert
+ * @returns {Blob} - The Blob
+ */
+export function dataURLtoBlob(dataurl) {
+    let arr = dataurl.split(","), mime = arr[0].match(/:(.*?);/)[1],
+        bstr = atob(arr[1]), n = bstr.length, u8arr = new Uint8Array(n);
+    while(n--){
+        u8arr[n] = bstr.charCodeAt(n);
+    }
+    return new Blob([u8arr], {type:mime});
+}
+
 export default {
     randomIntFromInterval,
     replaceAll,